requirements.txt
asgiref==3.7.2
cffi==1.15.1
channels==4.0.0
colorama==0.4.6
cryptography==41.0.2
Django==4.2.2
djangorestframework==3.14.0
gunicorn==21.2.0
# numpy==1.25.2
opencv-python==4.8.0.76
packaging==23.2
Pillow==10.0.0
pycparser==2.21
pyodbc==4.0.39
pypng==0.20220715.0
pytz==2023.3
qrcode==7.4.2
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
